move post into (tabs)
