--- frontend/app/(app)/(tabs)/post.tsx
@@ -1,12 +1,0 @@
-import { router } from 'expo-router';
-import { useEffect } from 'react';
-
-const Post = () => {
-  useEffect(() => {
-    router.replace('/(postcreation)');
-  }, []);
-
-  return null;
-};
-
-export default Post;
