+++ frontend/app/(app)/(tabs)/post/_layout.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { router, Stack } from 'expo-router';
 import { useForm, FormProvider } from 'react-hook-form';
-import Arrow from '../../../components/arrow';
-import { FormFields } from '../../../types/divelog';
+import Arrow from '../../../../components/arrow';
+import { FormFields } from '../../../../types/divelog';
 
 export default function Layout() {
   const methods = useForm<FormFields>();
@@ -11,7 +11,7 @@ export default function Layout() {
     <FormProvider {...methods}>
       <Stack>
         <Stack.Screen
-          name="post"
+          name="index"
           options={{
             headerTitle: '',
             headerTransparent: true,
