+++ frontend/app/(app)/(tabs)/post/index.tsx
@@ -5,7 +5,10 @@ import PostCreationForm from './components/post-creation-form';
 export default function PostCreation() {
   return (
     <View className="flex items-center w-full bg-white h-full pt-[5vh]">
-      <ScrollView className="flex flex-col w-[90vw] bg-white">
+      <ScrollView
+        showsVerticalScrollIndicator={false}
+        className="flex flex-col w-[90vw] bg-white"
+      >
         <View className="flex flex-col justify-items items-center pt-[5vh] mb-[3vh]">
           <Text className="font-bold text-[24px] leading-[29.05px] pb-[1vh]">
             {' '}
