+++ frontend/app/(app)/_layout.tsx
@@ -6,7 +6,6 @@ const Layout = () => {
     <Stack>
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
       <Stack.Screen name="user" options={{ headerShown: false }} />
-      <Stack.Screen name="(postcreation)" options={{ headerShown: false }} />
       <Stack.Screen name="divelog" options={{ headerShown: false }} />
     </Stack>
   );
